requirements.txt
Django==4.2.27
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-redis==5.4.0
celery==5.3.4
hiredis==2.2.3
redis==5.0.1
django-cors-headers==4.3.1
